removed new example app to make this pr alot smaller. Will add the new example app in a new pr
